+++ src/components/step-flow/index.ts
@@ -1,2 +1,7 @@
 export { default as StepFlow } from "./step-flow.component";
-export type { StepFlowProps } from "./step-flow.component";
+export { default as StepFlowTitle } from "./step-flow-title/step-flow-title.component";
+export type {
+  StepFlowProps,
+  StepFlowHandle,
+  Steps,
+} from "./step-flow.component";